Add logging, run sanitizers
AddrSan, LeakSan, and UBSan all come back clean for all demos.
There is an issue where MemSan freaks out with ostream's and
a uint32_t being unitialized, but that may be due to a lack of
library support.

There is now a logger. It is slow bc it is directly using ostream's
with no buffer but it works. We'll add a buffer in the main project.

Co-authored-by: Dov Kruger <hydrodog@users.noreply.github.com>